+++ src/index.ts
@@ -1,8 +1,12 @@
 /**
- * Copyright (C) 2020 IKOA Business Opportunity
+ * Copyright (C) 2020 - 2021 IKOA Business Opportunity
+ *
+ * All Rights Reserved
  * Author: Reinier Millo Sánchez <millo@ikoabo.com>
  *
- * This file is part of the IKOA Business Opportunity Auth API.
+ * This file is part of the IKOA Business Oportunity Auth Package
+ * It can't be copied and/or distributed without the express
+ * permission of the author.
  */
 
 /* Export api middlewares */