+++ CHANGELOG.md
@@ -1,15 +1,32 @@
 # CHANGELOG
 
+## [1.1.0] - 2021-07-17
+
+- Update errors definition
+- Fix url to login
+- Handle project in auth urls
+
+## [1.0.19] - 2020-10-19
+
+- Adding query bearer token authentication if no header present
+
+## [1.0.18] - 2020-10-13
+
+- Return username on authentication validation
+
 ## [1.0.17] - 2020-09-04
+
 - Add middleware to allow pass access token into query parameters
 
 ## [1.0.16] - 2020-09-03
+
 - Fix scope AND validation
 - Replace request dependencie by axios
 - Add axios interceptor
 - Update dependencies
 
 ## [1.0.14] - 2020-08-23
+
 - Initial public version
 - Predefined authentication error constants
 - Authentication middleware